refactor: 에피소드 페이지의 UI 이동 완료 (#99)
