+++ apps/web/src/features/episode/EpisodeDetailSection.tsx
@@ -2,7 +2,7 @@ import { Card } from '@repo/ui/components/ui/card';
 import { UserInfoType } from 'src/app/api/auth/user-info/route';
 import { fetchEpisodesById } from 'src/entities/episode/api';
 import { AuctionRow, SellerRow } from 'src/shared/supabase/types';
-import EpisodeList from './EpisodeList';
+import EpisodeList from 'src/features/episode/EpisodeList';
 
 const EpisodeDetailSection = async ({
   auctionId,
